changed item price to double to avoid merge conflicts later

+++ src/main/java/com/qa/ims/persistence/dao/ItemDaoMysql.java
@@ -36,7 +36,7 @@ public class ItemDaoMysql implements Dao<Item> {
 	Item itemFromResultSet(ResultSet resultSet) throws SQLException {
 		Long itemID = resultSet.getLong("itemID");
 		String itemName = resultSet.getString("itemName");
-		int itemPrice = resultSet.getInt("itemPrice");
+		Double itemPrice = resultSet.getDouble("itemPrice");
 		return new Item(itemID, itemName, itemPrice);
 
 	}
